Move timezones to npm package and convert util-frontend.js to typescript
